ref(profiledetails): split to separate component
