+++ static/app/components/profiling/FrameStack/frameStack.tsx
@@ -1,9 +1,7 @@
 import {memo, MouseEventHandler, useCallback, useMemo, useState} from 'react';
 import styled from '@emotion/styled';
-import {PlatformIcon} from 'platformicons';
 
 import Button from 'sentry/components/button';
-import DateTime from 'sentry/components/dateTime';
 import {ExportProfileButton} from 'sentry/components/profiling/exportProfileButton';
 import {IconPanel} from 'sentry/icons';
 import {t} from 'sentry/locale';
@@ -16,10 +14,10 @@ import {useFlamegraphPreferences} from 'sentry/utils/profiling/flamegraph/useFla
 import {FlamegraphFrame} from 'sentry/utils/profiling/flamegraphFrame';
 import {ProfileGroup} from 'sentry/utils/profiling/profile/importProfile';
 import {invertCallTree} from 'sentry/utils/profiling/profile/utils';
-import {makeFormatter} from 'sentry/utils/profiling/units/units';
 import {useParams} from 'sentry/utils/useParams';
 
 import {FrameStackTable} from './frameStackTable';
+import {ProfileDetails} from './profileDetails';
 
 interface FrameStackProps {
   canvasPoolManager: CanvasPoolManager;
@@ -38,7 +36,6 @@ const FrameStack = memo(function FrameStack(props: FrameStackProps) {
     useFlamegraphPreferences();
 
   const [tab, setTab] = useState<'bottom up' | 'call order'>('call order');
-  const [detailsTab, setDetailsTab] = useState<'device' | 'transaction'>('transaction');
   const [treeType, setTreeType] = useState<'all' | 'application' | 'system'>('all');
   const [recursion, setRecursion] = useState<'collapsed' | null>(null);
 
@@ -89,14 +86,6 @@ const FrameStack = memo(function FrameStack(props: FrameStackProps) {
     setTreeType('system');
   }, []);
 
-  const onDeviceTabClick = useCallback(() => {
-    setDetailsTab('device');
-  }, []);
-
-  const onTransactionTabClick = useCallback(() => {
-    setDetailsTab('transaction');
-  }, []);
-
   const onTableLeftClick = useCallback(() => {
     dispatchFlamegraphPreferences({type: 'set layout', payload: 'table left'});
   }, [dispatchFlamegraphPreferences]);
@@ -111,8 +100,8 @@ const FrameStack = memo(function FrameStack(props: FrameStackProps) {
 
   return (
     <FrameDrawer layout={flamegraphPreferences.layout}>
-      <FrameTabs>
-        <ListItem className={tab === 'bottom up' ? 'active' : undefined}>
+      <ProfilingDetailsFrameTabs>
+        <ProfilingDetailsListItem className={tab === 'bottom up' ? 'active' : undefined}>
           <Button
             data-title={t('Bottom Up')}
             priority="link"
@@ -121,8 +110,11 @@ const FrameStack = memo(function FrameStack(props: FrameStackProps) {
           >
             {t('Bottom Up')}
           </Button>
-        </ListItem>
-        <ListItem margin="none" className={tab === 'call order' ? 'active' : undefined}>
+        </ProfilingDetailsListItem>
+        <ProfilingDetailsListItem
+          margin="none"
+          className={tab === 'call order' ? 'active' : undefined}
+        >
           <Button
             data-title={t('Call Order')}
             priority="link"
@@ -131,9 +123,9 @@ const FrameStack = memo(function FrameStack(props: FrameStackProps) {
           >
             {t('Call Order')}
           </Button>
-        </ListItem>
+        </ProfilingDetailsListItem>
         <Separator />
-        <ListItem className={treeType === 'all' ? 'active' : undefined}>
+        <ProfilingDetailsListItem className={treeType === 'all' ? 'active' : undefined}>
           <Button
             data-title={t('All Frames')}
             priority="link"
@@ -142,8 +134,10 @@ const FrameStack = memo(function FrameStack(props: FrameStackProps) {
           >
             {t('All Frames')}
           </Button>
-        </ListItem>
-        <ListItem className={treeType === 'application' ? 'active' : undefined}>
+        </ProfilingDetailsListItem>
+        <ProfilingDetailsListItem
+          className={treeType === 'application' ? 'active' : undefined}
+        >
           <Button
             data-title={t('Application Frames')}
             priority="link"
@@ -152,8 +146,11 @@ const FrameStack = memo(function FrameStack(props: FrameStackProps) {
           >
             {t('Application Frames')}
           </Button>
-        </ListItem>
-        <ListItem margin="none" className={treeType === 'system' ? 'active' : undefined}>
+        </ProfilingDetailsListItem>
+        <ProfilingDetailsListItem
+          margin="none"
+          className={treeType === 'system' ? 'active' : undefined}
+        >
           <Button
             data-title={t('System Frames')}
             priority="link"
@@ -162,9 +159,9 @@ const FrameStack = memo(function FrameStack(props: FrameStackProps) {
           >
             {t('System Frames')}
           </Button>
-        </ListItem>
+        </ProfilingDetailsListItem>
         <Separator />
-        <ListItem>
+        <ProfilingDetailsListItem>
           <FrameDrawerLabel>
             <input
               type="checkbox"
@@ -173,8 +170,8 @@ const FrameStack = memo(function FrameStack(props: FrameStackProps) {
             />
             {t('Collapse recursion')}
           </FrameDrawerLabel>
-        </ListItem>
-        <ListItem
+        </ProfilingDetailsListItem>
+        <ProfilingDetailsListItem
           style={{
             flex: '1 1 100%',
             cursor:
@@ -184,7 +181,7 @@ const FrameStack = memo(function FrameStack(props: FrameStackProps) {
             flamegraphPreferences.layout === 'table bottom' ? props.onResize : undefined
           }
         />
-        <ListItem margin="none">
+        <ProfilingDetailsListItem margin="none">
           <ExportProfileButton
             variant="xs"
             eventId={params.eventId}
@@ -196,9 +193,9 @@ const FrameStack = memo(function FrameStack(props: FrameStackProps) {
               params.projectId === undefined
             }
           />
-        </ListItem>
+        </ProfilingDetailsListItem>
         <Separator />
-        <ListItem>
+        <ProfilingDetailsListItem>
           <LayoutSelectionContainer>
             <StyledButton
               active={flamegraphPreferences.layout === 'table left'}
@@ -225,8 +222,8 @@ const FrameStack = memo(function FrameStack(props: FrameStackProps) {
               <IconPanel size="xs" direction="left" />
             </StyledButton>
           </LayoutSelectionContainer>
-        </ListItem>
-      </FrameTabs>
+        </ProfilingDetailsListItem>
+      </ProfilingDetailsFrameTabs>
 
       <FrameStackTable
         {...props}
@@ -235,86 +232,8 @@ const FrameStack = memo(function FrameStack(props: FrameStackProps) {
         tree={maybeFilteredOrInvertedTree ?? []}
         canvasPoolManager={props.canvasPoolManager}
       />
-      <ProfileDetails layout={flamegraphPreferences.layout}>
-        <FrameTabs>
-          <ListItem
-            size="sm"
-            className={detailsTab === 'transaction' ? 'active' : undefined}
-          >
-            <Button
-              data-title={t('Transaction')}
-              priority="link"
-              size="zero"
-              onClick={onTransactionTabClick}
-            >
-              {t('Transaction')}
-            </Button>
-          </ListItem>
-          <ListItem size="sm" className={detailsTab === 'device' ? 'active' : undefined}>
-            <Button
-              data-title={t('Device')}
-              priority="link"
-              size="zero"
-              onClick={onDeviceTabClick}
-            >
-              {t('Device')}
-            </Button>
-          </ListItem>
-        </FrameTabs>
 
-        {detailsTab === 'device' ? (
-          <DetailsContainer>
-            <DetailsRow>
-              {Object.entries(DEVICE_DETAILS_KEY).map(([label, key]) => {
-                const value = props.profileGroup.metadata[key];
-                return (
-                  <DetailsRow key={key}>
-                    <span>
-                      <strong>{label}</strong>:
-                    </span>{' '}
-                    <span>
-                      {key === 'durationNS'
-                        ? nsFormatter(value)
-                        : value === undefined || value === ''
-                        ? t('ø')
-                        : value}
-                    </span>
-                  </DetailsRow>
-                );
-              })}
-            </DetailsRow>
-          </DetailsContainer>
-        ) : (
-          <DetailsContainer>
-            {Object.entries(PROFILE_DETAILS_KEY).map(([label, key]) => {
-              const value = props.profileGroup.metadata[key];
-
-              return (
-                <DetailsRow key={key}>
-                  <strong>{label}</strong>:{' '}
-                  <span>
-                    {key === 'durationNS' ? (
-                      nsFormatter(value)
-                    ) : key === 'threads' ? (
-                      props.profileGroup.profiles.length
-                    ) : key === 'received' ? (
-                      <DateTime date={value} />
-                    ) : value === undefined || value === '' ? (
-                      t('ø')
-                    ) : (
-                      value
-                    )}
-                    {key === 'platform' ? (
-                      <PlatformIcon size={12} platform={value ?? 'unknown'} />
-                    ) : null}
-                  </span>
-                </DetailsRow>
-              );
-            })}
-            <DetailsRow />
-          </DetailsContainer>
-        )}
-      </ProfileDetails>
+      <ProfileDetails profileGroup={props.profileGroup} />
 
       {flamegraphPreferences.layout === 'table left' ||
       flamegraphPreferences.layout === 'table right' ? (
@@ -327,62 +246,6 @@ const FrameStack = memo(function FrameStack(props: FrameStackProps) {
     </FrameDrawer>
   );
 });
-
-const nsFormatter = makeFormatter('nanoseconds');
-
-const PROFILE_DETAILS_KEY: Record<string, string> = {
-  [t('transaction')]: 'transactionName',
-  [t('received at')]: 'received',
-  [t('organization')]: 'organizationID',
-  [t('project')]: 'projectID',
-  [t('platform')]: 'platform',
-  [t('environment')]: 'environment',
-  [t('version')]: 'version',
-  [t('duration')]: 'durationNS',
-  [t('threads')]: 'threads',
-};
-
-const DEVICE_DETAILS_KEY: Record<string, string> = {
-  [t('model')]: 'deviceModel',
-  [t('manufacturer')]: 'deviceManufacturer',
-  [t('classification')]: 'deviceClassification',
-  [t('os')]: 'deviceOSName',
-  [t('os version')]: 'deviceOSVersion',
-  [t('locale')]: 'deviceLocale',
-};
-
-const DetailsRow = styled('div')`
-  white-space: nowrap;
-  text-overflow: ellipsis;
-  overflow: hidden;
-  font-size: ${p => p.theme.fontSizeSmall};
-`;
-
-const DetailsContainer = styled('ul')`
-  padding: ${space(1)};
-  margin: 0;
-  overflow: auto;
-  position: absolute;
-  left: 0;
-  top: 24px;
-  width: 100%;
-  height: calc(100% - 24px);
-`;
-
-const ProfileDetails = styled('div')<{layout: FlamegraphPreferences['layout']}>`
-  width: ${p =>
-    p.layout === 'table left' || p.layout === 'table right' ? '100%' : '260px'};
-  height: ${p =>
-    p.layout === 'table left' || p.layout === 'table right' ? '220px' : '100%'};
-  border-left: 1px solid ${p => p.theme.border};
-  background: ${p => p.theme.background};
-  grid-area: details;
-  position: relative;
-
-  > ul:first-child {
-    border-bottom: 1px solid ${p => p.theme.border};
-  }
-`;
 
 const ResizableVerticalDrawer = styled('div')`
   width: 1px;
@@ -450,7 +313,7 @@ const Separator = styled('li')`
   transform: translateY(29%);
 `;
 
-const FrameTabs = styled('ul')`
+export const ProfilingDetailsFrameTabs = styled('ul')`
   display: flex;
   list-style-type: none;
   padding: 0 ${space(1)};
@@ -461,7 +324,10 @@ const FrameTabs = styled('ul')`
   grid-area: tabs;
 `;
 
-const ListItem = styled('li')<{margin?: 'none'; size?: 'sm'}>`
+export const ProfilingDetailsListItem = styled('li')<{
+  margin?: 'none';
+  size?: 'sm';
+}>`
   font-size: ${p => p.theme.fontSizeSmall};
   margin-right: ${p => (p.margin === 'none' ? 0 : space(1))};
