+++ static/app/components/profiling/FrameStack/profileDetails.tsx
@@ -1,0 +1,174 @@
+import {useCallback, useState} from 'react';
+import styled from '@emotion/styled';
+import {PlatformIcon} from 'platformicons';
+
+import Button from 'sentry/components/button';
+import DateTime from 'sentry/components/dateTime';
+import {t} from 'sentry/locale';
+import space from 'sentry/styles/space';
+import {FlamegraphPreferences} from 'sentry/utils/profiling/flamegraph/flamegraphStateProvider/flamegraphPreferences';
+import {useFlamegraphPreferencesValue} from 'sentry/utils/profiling/flamegraph/useFlamegraphPreferences';
+import {ProfileGroup} from 'sentry/utils/profiling/profile/importProfile';
+import {makeFormatter} from 'sentry/utils/profiling/units/units';
+
+import {ProfilingDetailsFrameTabs, ProfilingDetailsListItem} from './frameStack';
+
+interface ProfileDetailsProps {
+  profileGroup: ProfileGroup;
+}
+
+export function ProfileDetails(props: ProfileDetailsProps) {
+  const [detailsTab, setDetailsTab] = useState<'device' | 'transaction'>('transaction');
+
+  const onDeviceTabClick = useCallback(() => {
+    setDetailsTab('device');
+  }, []);
+
+  const onTransactionTabClick = useCallback(() => {
+    setDetailsTab('transaction');
+  }, []);
+
+  const flamegraphPreferences = useFlamegraphPreferencesValue();
+
+  return (
+    <ProfileDetailsBar layout={flamegraphPreferences.layout}>
+      <ProfilingDetailsFrameTabs>
+        <ProfilingDetailsListItem
+          size="sm"
+          className={detailsTab === 'transaction' ? 'active' : undefined}
+        >
+          <Button
+            data-title={t('Transaction')}
+            priority="link"
+            size="zero"
+            onClick={onTransactionTabClick}
+          >
+            {t('Transaction')}
+          </Button>
+        </ProfilingDetailsListItem>
+        <ProfilingDetailsListItem
+          size="sm"
+          className={detailsTab === 'device' ? 'active' : undefined}
+        >
+          <Button
+            data-title={t('Device')}
+            priority="link"
+            size="zero"
+            onClick={onDeviceTabClick}
+          >
+            {t('Device')}
+          </Button>
+        </ProfilingDetailsListItem>
+      </ProfilingDetailsFrameTabs>
+
+      {detailsTab === 'device' ? (
+        <DetailsContainer>
+          <DetailsRow>
+            {Object.entries(DEVICE_DETAILS_KEY).map(([label, key]) => {
+              const value = props.profileGroup.metadata[key];
+              return (
+                <DetailsRow key={key}>
+                  <span>
+                    <strong>{label}</strong>:
+                  </span>{' '}
+                  <span>
+                    {key === 'durationNS'
+                      ? nsFormatter(value)
+                      : value === undefined || value === ''
+                      ? t('ø')
+                      : value}
+                  </span>
+                </DetailsRow>
+              );
+            })}
+          </DetailsRow>
+        </DetailsContainer>
+      ) : (
+        <DetailsContainer>
+          {Object.entries(PROFILE_DETAILS_KEY).map(([label, key]) => {
+            const value = props.profileGroup.metadata[key];
+
+            return (
+              <DetailsRow key={key}>
+                <strong>{label}</strong>:{' '}
+                <span>
+                  {key === 'durationNS' ? (
+                    nsFormatter(value)
+                  ) : key === 'threads' ? (
+                    props.profileGroup.profiles.length
+                  ) : key === 'received' ? (
+                    <DateTime date={value} />
+                  ) : value === undefined || value === '' ? (
+                    t('ø')
+                  ) : (
+                    value
+                  )}
+                  {key === 'platform' ? (
+                    <PlatformIcon size={12} platform={value ?? 'unknown'} />
+                  ) : null}
+                </span>
+              </DetailsRow>
+            );
+          })}
+          <DetailsRow />
+        </DetailsContainer>
+      )}
+    </ProfileDetailsBar>
+  );
+}
+
+const nsFormatter = makeFormatter('nanoseconds');
+
+const PROFILE_DETAILS_KEY: Record<string, string> = {
+  [t('transaction')]: 'transactionName',
+  [t('received at')]: 'received',
+  [t('organization')]: 'organizationID',
+  [t('project')]: 'projectID',
+  [t('platform')]: 'platform',
+  [t('environment')]: 'environment',
+  [t('version')]: 'version',
+  [t('duration')]: 'durationNS',
+  [t('threads')]: 'threads',
+};
+
+const DEVICE_DETAILS_KEY: Record<string, string> = {
+  [t('model')]: 'deviceModel',
+  [t('manufacturer')]: 'deviceManufacturer',
+  [t('classification')]: 'deviceClassification',
+  [t('os')]: 'deviceOSName',
+  [t('os version')]: 'deviceOSVersion',
+  [t('locale')]: 'deviceLocale',
+};
+
+const DetailsRow = styled('div')`
+  white-space: nowrap;
+  text-overflow: ellipsis;
+  overflow: hidden;
+  font-size: ${p => p.theme.fontSizeSmall};
+`;
+
+const DetailsContainer = styled('ul')`
+  padding: ${space(1)};
+  margin: 0;
+  overflow: auto;
+  position: absolute;
+  left: 0;
+  top: 24px;
+  width: 100%;
+  height: calc(100% - 24px);
+`;
+
+const ProfileDetailsBar = styled('div')<{layout: FlamegraphPreferences['layout']}>`
+  width: ${p =>
+    p.layout === 'table left' || p.layout === 'table right' ? '100%' : '260px'};
+  height: ${p =>
+    p.layout === 'table left' || p.layout === 'table right' ? '220px' : '100%'};
+  border-left: 1px solid ${p => p.theme.border};
+  background: ${p => p.theme.background};
+  grid-area: details;
+  position: relative;
+
+  > ul:first-child {
+    border-bottom: 1px solid ${p => p.theme.border};
+  }
+`;
